Make the install script understand muddle build trees

+++ sandbox/ToDoFinder/install.py
@@ -2,7 +2,8 @@
 
 """Install the TODO and //== pre-commit hook.
 
-Usage: install.py  [-f]  [<target> [<target> [...]]
+Usage: install.py  [-f] [-n]  [<target> [<target> [...]]
+       install.py  [-f] [-n]  -muddle [<muddle-root>]
 
 For each <target> (default is the current directory), look in that directory
 for a .git directory, and install the pre-commit hook therein.
@@ -11,11 +12,21 @@ Refuses if there already is a pre-commit hook, unless -f is specified.
 
 (nb: if -f is not specified, checks all the <target> directories before it does
 anything).
+
+If -n is specified, say what we would do, but don't do it.
+
+(If both -n and -f are specified, then -n "wins")
+
+Alternatively, specify -muddle to insert a pre-commit hook in each .git
+directory of each checkout in the current muddle tree (or the muddle tree
+at <muddle-root> if that is specified). This requires that the muddle command
+be "muddle".
 """
 
 import os
 import sys
 import shutil
+import subprocess
 
 def normalise_path(path):
     path = os.path.expanduser(path)
@@ -27,6 +38,10 @@ def main(args):
 
     targets = []
     force = False
+    dry_run = False
+    muddle = False
+    muddle_root = None
+
     while args:
 
         word = args.pop(0)
@@ -35,8 +50,90 @@ def main(args):
             return
         elif word == '-f':
             force = True
+        elif word == '-n':
+            dry_run = True
+        elif word == '-muddle':
+            muddle = True
+            # I don't like making -muddle have to be last on the command line,
+            # but it's the simplest thing to do...
+            if args:
+                if len(args) != 1:
+                    print 'Too many arguments after -muddle'
+                    return 1
+                muddle_root = args[0]
+                break
         else:
             targets.append(normalise_path(word))
+
+    this_file = __file__
+    this_dir = normalise_path(os.path.split(this_file)[0])
+
+    if muddle:
+        # Rather horribly, make sure we use the muddle we came with
+        our_muddle = normalise_path(os.path.join(this_dir, '..', '..', 'muddle'))
+        if not muddle_root:
+            # We use shell=True in case "muddle" is (for instance) an alias
+            # rather than on the path
+            try:
+                cmd = [our_muddle, 'query', 'root']
+                muddle_root = subprocess.check_output(cmd)
+            except subprocess.CalledProcessError as e:
+                print '%r returned %d and said:'%(' '.join(cmd), e.returncode)
+                print e.output
+                return 1
+            muddle_root.strip()
+        print 'muddle_root', muddle_root
+
+        cmd = [our_muddle, '--tree', muddle_root, 'query', 'checkout-dirs']
+        try:
+            checkout_report = subprocess.check_output(cmd)
+        except subprocess.CalledProcessError as e:
+            print '%r returned %d and said:'%(' '.join(cmd), e.returncode)
+            print e.output
+            return 1
+
+        checkout_paths = {}
+        lines = checkout_report.split('\n')
+        for line in lines:
+            if not line or line.startswith('> Checkout'):
+                continue
+            words = line.split()
+            checkout = words[0]
+            rest = line[len(checkout):].lstrip()
+            if rest.startswith('->'):
+                path = rest[len('->'):].lstrip()
+            else:
+                print 'Expected "->" in %r'%line
+                return 1
+            checkout_paths[checkout] = path
+
+        cmd = [our_muddle, '--tree', muddle_root, 'query', 'checkout-vcs']
+        try:
+            checkout_report = subprocess.check_output(cmd)
+        except subprocess.CalledProcessError as e:
+            print '%r returned %d and said:'%(' '.join(cmd), e.returncode)
+            print e.output
+            return 1
+
+        git_checkouts = []
+        lines = checkout_report.split('\n')
+        for line in lines:
+            if not line or line.startswith('> Checkout'):
+                continue
+            words = line.split()
+            checkout = words[0]
+            vcs = words[-1]
+            if vcs == 'git':
+                git_checkouts.append(checkout)
+            else:
+                print 'Ignoring %s checkout %s'%(vcs, checkout)
+
+        for checkout in git_checkouts:
+            targets.append(os.path.join(muddle_root, checkout_paths[checkout]))
+
+        if not targets:
+            print 'No git checkouts in %s'%muddle_root
+            return 1
 
     if not targets:
         targets.append(normalise_path('.'))
@@ -69,8 +166,6 @@ def main(args):
             print 'Giving up'
             return 1
 
-    this_file = __file__
-    this_dir = normalise_path(os.path.split(this_file)[0])
     pre_commit_file = os.path.join(this_dir, 'pre-commit')
     check_names_file = os.path.join(this_dir, 'FullCheckNames.py')
 
@@ -78,8 +173,12 @@ def main(args):
     print check_names_file
 
     for hook_dir in hook_dirs:
-        shutil.copy2(pre_commit_file, hook_dir)
-        shutil.copy2(check_names_file, hook_dir)
+        if dry_run:
+            print 'Copy %s to %s'%(pre_commit_file, hook_dir)
+            print ' and %s to %s'%(check_names_file, hook_dir)
+        else:
+            shutil.copy2(pre_commit_file, hook_dir)
+            shutil.copy2(check_names_file, hook_dir)
 
 if __name__ == '__main__':
     args = sys.argv[1:]
